pyproject.toml
[project]
name = "medichat"
version = "0.1.0"
description = ""
authors = [
    {name = "mal-mak",email = "malekhlouf@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "streamlit (>=1.42.2,<2.0.0)",
    "google-cloud-storage (>=3.0.0,<4.0.0)",
    "python-magic (>=0.4.27,<0.5.0)",
    "sqlalchemy (>=2.0.38,<3.0.0)",
    "langchain-google-cloud-sql-pg (>=0.12.1,<0.13.0)",
    "langchain-core (>=0.3.40,<0.4.0)",
    "langchain (>=0.3.19,<0.4.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "uvicorn (>=0.34.0,<0.35.0)"
]

[tool.poetry]
packages = [{include = "medichat", from = "src"}]


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
